broke the huge dashboard into separate dashboards finally, adjusted routing, fixed some loop keys and missing translations
