+++ client/src/constants/styles.ts
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+
+export const DashboardTitle = styled.h2`
+    font-weight: 600;
+`;
+
+export const DashboardSubTitle = styled.h3`
+    font-weight: 600;
+`;
+
+export const DashboardGridColumnHeader = styled.span`
+    display: flex;
+    justify-content: center;
+`;
+
